#1315 longer delay

+++ src/main/resources/com/rultor/agents/docker/rmi.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 set -e
 
-images=$(docker images | grep -E '([0-9]{2,} weeks|[0-9]+ months) ago' | tr -s ' ' | cut -d ' ' -f 3 | uniq)
+images=$(docker images --no-trunc | grep -E '([0-9]{2,} weeks|[0-9]+ months) ago' | tr -s ' ' | cut -d ' ' -f 3 | uniq)
 for i in ${images}; do
-  docker rmi -f "${i}" || echo "Failed to remove Docker image ${i}"
+  containers=$(docker ps --filter="ancestor=${i}" | wc -l)
+  if [ "${containers}" -eq "1" ]; then
+    # echo "Docker image ${i} is being used, won't delete"
+  else
+      docker rmi -f "${i}" || echo "Failed to remove Docker image ${i}"
+  fi
 done
